% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{coef2}
\alias{coef2}
\title{An alternative generic to coef}
\usage{
coef2(object, ...)
}
\arguments{
\item{object}{a gfo object}

\item{...}{additional arguments}
}
\description{
This was created for internal purposes, to create a common interface between
different types of modeling output
}
