% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ComputeMcFaddensR2}
\alias{ComputeMcFaddensR2}
\title{Computes the Cox & Snell R-squared}
\usage{
ComputeMcFaddensR2(fit, null.model)
}
\arguments{
\item{fit}{a full model object, must support the LogLik function}

\item{null.model}{the reduced model object, must support the LogLike function}

\item{n}{the sample size}
}
\value{
the scalar Cox & Snell R-squared
}
\description{
Computes the Cox & Snell R-squared
}

