% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CreateFormula}
\alias{CreateFormula}
\title{Formula creation from string arguments}
\usage{
CreateFormula(response, predictors, random.terms = character(0))
}
\arguments{
\item{response}{the response variable's name}

\item{predictors}{a character vector of predictor names}

\item{random.terms}{a character vector of random terms as they are presented in
terms(formula)}
}
\description{
Formula creation from string arguments
}
