% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential.R
\name{GetFastFSR}
\alias{GetFastFSR}
\title{INCOMPLETE - Compute Fast FSR estimate}
\usage{
GetFastFSR(n.total.vars, n.model.vars, alpha, verbose = FALSE)
}
\arguments{
\item{n.total.vars}{Number of variables eligible for model}

\item{n.model.vars}{Number of variables kept in model}

\item{alpha}{the traditional alpha-to-enter or alpha-to-stay}

\item{verbose}{toggles additional output}
}
\value{
The FSR estimate of Boos, Stefanski, and Wu (2009)
}
\description{
http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2748177/
}
