% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfo_assessment.R
\name{GetVariableImpacts}
\alias{GetVariableImpacts}
\title{GetVariableImpacts}
\usage{
GetVariableImpacts(obj, lower.quantile = 0.05, upper.quantile = 0.95)
}
\arguments{
\item{obj}{either a gfo.cxv (cross validation) or gfo object}

\item{lower.quantile}{the starting quantile of each continuous predictor}

\item{upper.quantile}{the ending quantile of each continuous predictor}
}
\value{
A data frame of the predicted outcomes under the modulations
}
\description{
Computes model extrapolations when fixed effects are modulated
}
\details{
Currently works with only a mids data object and family = binomial!
}
\examples{
data(testdata)
mids <- ImputeData(testdata, m = 5, maxit = 5)
gfo <- BackwardEliminate(y ~ x + w + z, data = mids)
cv <- CrossValidate(gfo)
impact <- GetVariableImpacts(cv)
impact

}

