% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{MakeSplineStr}
\alias{MakeSplineStr}
\title{Make a spline formula string corresponding to unambiguous spline}
\usage{
MakeSplineStr(var.name, data, interior.knots)
}
\arguments{
\item{var.name}{name of variable for use with spline equation}

\item{data}{data set that will be used to construct the spline string}

\item{interior.knots}{the interior knots of the spline}
}
\description{
The ns() function is convenient, but without specifiying the knots and the
boundary knots, the information neccessary to reproduce the function is in
the data set itself. This function creates a string that can be used in
future formulas and does not depend on the original data set for
reproducibility.
}

