% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\docType{data}
\name{nls.97}
\alias{nls.97}
\title{Sample from the National Longitudinal Survey of Youth - NLSY97 (1997-2013)}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(nls.97)
}
\description{
A dataset compiled using the NLS Investigator online tool
(www.nlsinfo.org/investigator) along with the NLSdata package
(www.github.com/google/NLSdata) with the goal of investigating the link
to a math or computer science career to attitudes and demographic variables.
}
\details{
\itemize{
  \item PUBID.1997. Subject level identifier in the NLSY97 survey
  \item age. Age of the respondent at time of answering questions
  \item math.cs. Whether current job is classified as math or CS
  \item gender. Self identified gender of the subject
  \item race.  Self identified race of the subject
  \item poverty.ratio. Ratio of household income to poverty level in the previous year 
  \item hs.prog.1997. Had subject taken high school programming class by the 1997 interview?
  \item teachers.good. 1-5, extent to which subject thought his or her teachers were good
  \item teachers.interested. 1-5, extent to which subject thought his or her teachers were interested
  \item enriching.environment. 0-3, higher scores indicate a more enriching environment 
  \item{prog.course.since.dli.1998.}{ When asked in 1998, did subject take
                                    programming course since data of last
                                    interview?}
  \item{prog.course.since.dli.2001.}{ When asked in 2001, did subject take
                                    programming course since data of last
                                    interview?}
  \item pr.five.yr.marriage. Subjects estimation of marriage probability in 5 years.
  \item organized. 1-5, self-rated, higher scores mean more organized
  \item conscientious. 1-5, self-rated, higher scores mean more conscientious 
  \item dependable. 1-5, self-rated, higher scores mean more dependable
  \item thorough. 1-5, self-rated, higher scores mean more thorough
  \item agreeable. 1-5, self-rated, higher scores mean more agreeable 
  \item cooperative. 1-5, self-rated, higher scores mean more cooperative 
  \item flexible. 1-5, self-rated, higher scores mean more flexible
  \item trustful. 1-5, self-rated, higher scores mean more trustful
}
}
\keyword{datasets}

