% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{plotFSR}
\alias{plotFSR}
\title{INCOMPLETE - function to plot FSR in fast method}
\usage{
plotFSR(gfo)
}
\arguments{
\item{gfo}{a gfo object}
}
\description{
INCOMPLETE - function to plot FSR in fast method
}
