% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{predict.gfo}
\alias{predict.gfo}
\title{predict.gfo}
\usage{
\method{predict}{gfo}(obj, newdata)
}
\arguments{
\item{obj}{A gfo object}

\item{newdata}{either a data.frame or a mids (multiply imputed data set) object}
}
\value{
numerical vector of fitted response values
}
\description{
Prediction function for a generalized fitted object from glmmplus package
}

