% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{scoef}
\alias{scoef}
\title{Standardized Coefficients}
\usage{
scoef(obj)
}
\arguments{
\item{obj}{a gfo object}
}
\description{
Computes the standardized coefficients from a gfo object
}

