% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata: a small data set with missing values}
\format{A data frame with 500 rows and 7 variables}
\usage{
data(testdata)
}
\description{
Contains numeric variables x (25% missing), y (25% missing), and w
(0% missing) as well as two factor variables (factor.1, factor.2) for use
with random effects modeling. Finally, y and y.binary are suggested for use
as example continuous and binary response variables, respectively.
}
\keyword{datasets}

