% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CreateFormula}
\alias{CreateFormula}
\title{Formula creation from string arguments}
\usage{
CreateFormula(response, predictors, random.terms = character(0))
}
\description{
@param response the response variable's name
 @param predictors a character vector of predictor names
 @param random.terms a character vector of random terms as they are presented in
               terms(formula)
}

