% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfo_assessment.R
\name{CrossValidate}
\alias{CrossValidate}
\title{CrossValidate}
\usage{
CrossValidate(gfo, k.folds = 5, granularity = 0.005, starter = list(),
  pre.test = FALSE)
}
\arguments{
\item{k.folds}{Number of cross-validation folds}

\item{granularity}{For binary classification, the step size from 0 to 1 in
computing evaluation metrics like precision}

\item{starter}{a list containing elements that allows restarting the
cross validation if at least one fold was completed.
Such a list is created at each iteration and placed in the
global environment. It is called "starter.pack"}

\item{pre.test}{Whether to run the full model for every training split,
to check for model stability before running a lengthy
cross-validation run using backward elimination}
}
\value{
A gfo.cxv object containing relevant cv outputs
}
\description{
Cross validation for use with gfo objects
}
\details{
Currently set up for classification, binomial family and mids data
         only!
}
\examples{
data(testdata)
creating a Muliply Imputed Data Set (mids) object
mids <- ImputeData(testdata, m = 5, maxit = 5)
gfo <- BackwardEliminate(y ~ x + w + z, data = mids)
cv <- CrossValidate(gfo)

}

