% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfo_assessment.R
\name{GetVariableGroupImpact}
\alias{GetVariableGroupImpact}
\title{GetVariableGroupImpact: Model Implied Group Impact Analysis}
\usage{
GetVariableGroupImpact(obj, low.val.list, high.val.list)
}
\arguments{
\item{obj}{A gfo object created by FitModel, BackwardEliminate, or
Forward Select}

\item{low.val.list}{A list mapping each variable name (from var.vec) to its
"low" value}

\item{high.val.list}{A list mapping each variable name (from var.vec) to its
"high" value}
}
\description{
Shows how the predicted values of a model change as a set of variables
simultaneously change in value.
}
\examples{
my.gfo <- FitModel(y ~ x + w + z, data = testdata)
group.impacts <- GetVariableGroupImpact(my.gfo,
                                        low.val.list = list(x = -1, w = -1),
                                        high.val.list = list(x = 1, w = 1))

}

