% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_extentions.R
\docType{import}
\name{reexports}
\alias{mice}
\alias{mice.impute.pmm}
\alias{pool}
\alias{pool.compare}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mice}{\code{\link[mice]{mice}}, \code{\link[mice]{mice.impute.pmm}}, \code{\link[mice]{pool}}, \code{\link[mice]{pool.compare}}}
}}
\keyword{internal}

