\name{scoef}
\alias{scoef}
\alias{Scoef}
\title{Standardized Coefficients}
\description{ This is ...}

\usage{
   scoef(gfo)
}

\arguments{
  \item{gfo}{
  A gfo object as returned by FitModel, BackwardEliminate,
  or ForwardSelect. 
  }
}
\details{
  This still needs to be tested against SPSS
}
\examples{
scoef(gfo.object)
}
